fix: 🦭 guard legacy subagent result parsing
