+++ src/components/chat/message/MessageBubble.tsx
@@ -48,6 +48,13 @@ import ModelPickerCard from "@/components/chat/ModelPickerCard"
 import ContextBreakdownCard from "@/components/chat/context-view/ContextBreakdownCard"
 import RecapProgressCard from "@/components/chat/RecapProgressCard"
 import SkillForkStatusCard from "@/components/chat/SkillForkStatusCard"
+import {
+  parseSubagentResultDetail,
+  parseSubagentResultStatus,
+  parseToolJobPayload,
+  TOOL_JOB_AGENT_PREFIX,
+  TOOL_JOB_STATUSES,
+} from "./asyncResultPayload"
 
 export interface MessageBubbleProps {
   msg: Message
@@ -84,102 +91,10 @@ export interface MessageBubbleProps {
   displayMode?: ChatDisplayMode
 }
 
-const TOOL_JOB_AGENT_PREFIX = "tool_job:"
-const TOOL_JOB_STATUSES = new Set([
-  "completed",
-  "failed",
-  "timed_out",
-  "cancelled",
-  "interrupted",
-  "running",
-])
-
 function hasRenderableTextContent(msg: Message): boolean {
   return (
     !!msg.content || !!msg.contentBlocks?.some((block) => block.type === "text" && !!block.content)
   )
-}
-
-function getXmlishAttribute(attrs: string, name: string): string | undefined {
-  const match = attrs.match(new RegExp(`\\b${name}="([^"]*)"`))
-  return match?.[1]
-}
-
-function getXmlishElement(content: string, name: string): string | undefined {
-  const match = content.match(new RegExp(`<${name}>([\\s\\S]*?)</${name}>`))
-  return match?.[1]?.trim()
-}
-
-function decodeXmlishText(value: string | undefined): string | undefined {
-  if (!value) return undefined
-  return value
-    .replace(/&lt;/g, "<")
-    .replace(/&gt;/g, ">")
-    .replace(/&amp;/g, "&")
-}
-
-function parseToolJobPayload(
-  content: string,
-): { toolName?: string; status?: string; detail?: string } | null {
-  const match = content.match(/<tool-job-result\b([^>]*)>/)
-  if (match) {
-    const attrs = match[1] || ""
-    return {
-      toolName: decodeXmlishText(getXmlishAttribute(attrs, "tool")),
-      status: decodeXmlishText(getXmlishAttribute(attrs, "status")),
-      detail:
-        decodeXmlishText(getXmlishElement(content, "output")) ||
-        decodeXmlishText(getXmlishElement(content, "error")) ||
-        decodeXmlishText(getXmlishElement(content, "note")),
-    }
-  }
-
-  if (!content.includes("<task-notification>")) {
-    return null
-  }
-  return {
-    toolName: decodeXmlishText(getXmlishElement(content, "tool")),
-    status: decodeXmlishText(getXmlishElement(content, "status")),
-    detail:
-      decodeXmlishText(getXmlishElement(content, "output-preview")) ||
-      decodeXmlishText(getXmlishElement(content, "error")) ||
-      decodeXmlishText(getXmlishElement(content, "summary")) ||
-      decodeXmlishText(getXmlishElement(content, "output-file")),
-  }
-}
-
-function parseSubagentResultDetail(content: string): string | undefined {
-  const xmlDetail =
-    decodeXmlishText(getXmlishElement(content, "result")) ||
-    decodeXmlishText(getXmlishElement(content, "error"))
-  if (xmlDetail) return xmlDetail
-
-  const match = content.match(
-    /<<<BEGIN_SUBAGENT_RESULT>>>\n?([\s\S]*?)\n?<<<END_SUBAGENT_RESULT>>>/,
-  )
-  return match?.[1]?.trim()
-}
-
-function parseSubagentResultStatus(content: string): string {
-  const status =
-    decodeXmlishText(getXmlishElement(content, "status")) ||
-    content.match(/^Status:\s*(\S+)/m)?.[1]
-  switch (status) {
-    case "completed":
-      return "completed"
-    case "timeout":
-    case "timed_out":
-      return "timed_out"
-    case "killed":
-      return "cancelled"
-    case "running":
-    case "spawning":
-      return "running"
-    case "error":
-      return "failed"
-    default:
-      return "completed"
-  }
 }
 
 function getSubagentResultDisplay(
